refactor: Refactor Operator
인터페이스와 클래스들로 구현했던 Operator를 Enum으로 바꾸고,
OperatorUtil을 추가해서 보다 클린한 코드로 리팩토링
