+++ src/main/java/calculator/Operator.java
@@ -1,22 +1,21 @@
 package calculator;
 
-public interface Operator {
-    static boolean isOperator(String value) {
-        try {
-            parseOperator(value);
-            return true;
-        } catch (IllegalArgumentException e) {
-            return false;
-        }
+import java.util.function.BiFunction;
+
+public enum Operator {
+    ADDITION((a, b) -> a + b),
+    SUBTRACTION((a, b) -> a - b),
+    MULTIPLICATION((a, b) -> a * b),
+    DIVISION((a, b) -> a / b);
+
+    final private BiFunction<Integer, Integer, Integer> calcFunc;
+
+    Operator(BiFunction<Integer, Integer, Integer> calcFunc) {
+        this.calcFunc = calcFunc;
     }
 
-    static Operator parseOperator(String value) {
-        if (value.equals("+")) return new Addition();
-        if (value.equals("-")) return new Subtraction();
-        if (value.equals("*")) return new Multiplication();
-        if (value.equals("/")) return new Division();
-        throw new IllegalArgumentException();
+    public int calculate(int a, int b) {
+        return calcFunc.apply(a, b);
     }
-
-    int calculate(int operand1, int operand2);
 }
+
