--- src/main/java/calculator/Subtraction.java
@@ -1,8 +1,0 @@
-package calculator;
-
-public class Subtraction implements Operator {
-    @Override
-    public int calculate(int operand1, int operand2) {
-        return operand1 - operand2;
-    }
-}
